+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/image/resizeable_image.dart
@@ -1,8 +1,6 @@
 import 'dart:io';
 import 'dart:math';
 
-import 'package:flutter/material.dart';
-
 import 'package:appflowy/generated/flowy_svgs.g.dart';
 import 'package:appflowy/plugins/document/application/prelude.dart';
 import 'package:appflowy/plugins/document/presentation/editor_plugins/image/common.dart';
@@ -10,6 +8,7 @@ import 'package:appflowy/shared/appflowy_network_image.dart';
 import 'package:appflowy_backend/protobuf/flowy-user/protobuf.dart';
 import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:string_validator/string_validator.dart';
@@ -61,7 +60,7 @@ class _ResizableImageState extends State<ResizableImage> {
   void initState() {
     super.initState();
     imageWidth = widget.width;
-    _userProfilePB = context.read<DocumentBloc>().state.userProfilePB;
+    _userProfilePB = context.read<DocumentBloc?>()?.state.userProfilePB;
   }
 
   @override
